drop [requests|builds|jobs].config unless on enterprise

+++ db/main/migrate/20180410000000_requests_drop_payload.rb
@@ -1,5 +1,5 @@
 class RequestsDropPayload < ActiveRecord::Migration[4.2]
   def change
-    remove_column :requests, :payload, :text
+    remove_column :requests, :payload, :text unless ENV['TRAVIS_ENTERPRISE']
   end
 end
